Modifies event listener and screen update functions
The event listener function now only makes the digit buttons
pass their value to the calculator. The variable which stores
said value was also modified and it's now an array instead of
a string.

The screen output fuction scans the array that stores the current
value. If the value stored at index 0 is 0, then it splices it before
updating the screen. Thus 0 never shows as the first digit of a number.